data_pipeline/car_traffic/requirements.txt
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
folium==0.20.0
geopandas==1.1.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
networkx==3.6
numpy==2.3.5
osmnx==2.0.7
packaging==25.0
pandas==2.3.3
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
shapely==2.1.2
six==1.17.0
tzdata==2025.2
urllib3==2.6.0
xyzservices==2025.11.0
